consolidates api-facing and regular modules into one location
