+++ routes/trains.js
@@ -1,5 +1,5 @@
 var express = require('express');
-var trains = require('../api_modules/mta_status')
+var trains = require('../modules/api-facing/mta_status')
 var router = express.Router();
 
 router.get('/',function(req, res) {
@@ -17,7 +17,7 @@ router.get('/:id', function(req, res) {
       var allTrains = status.map(function(train) {
         return train.name;
       });
-    
+
       // this is to keep the SIR as a single line
       var sir = allTrains.pop();
       allTrains = allTrains.join('').split('');
@@ -25,7 +25,7 @@ router.get('/:id', function(req, res) {
 
       res.json({status: allTrains});
     } else if(trains.getTrainString().indexOf(req.params.id) > -1){
-      
+
       var trainLine = status.filter(function(train) {
         if(train.name[0].indexOf(req.params.id) > -1) {
           return true;
@@ -33,7 +33,7 @@ router.get('/:id', function(req, res) {
           return false;
         }
       });
-      
+
       res.json({status: trainLine[0].status[0]});
     } else {
       res.json({status: 'error: no such train exists'});
